resolved issue in Section controller

+++ controllers/Section.js
@@ -1,5 +1,6 @@
 const Section = require('../models/Section');
-const Cource = require('../models/Course');
+const SubSection = require('../models/SubSection')
+const Course = require('../models/Course');
 
 exports.createSection = async (req,res) =>{
     try {
@@ -13,7 +14,7 @@ exports.createSection = async (req,res) =>{
         const newSection = await Section.create({sectionName});
 
         //upadating course with new section
-        const updatedCourse = await Cource.findByIdAndUpdate(courseId,
+        const updatedCourse = await Course.findByIdAndUpdate(courseId,
             {
                $push:{courseContent:newSection._id} 
             },
@@ -70,22 +71,48 @@ exports.updateSection = async (req,res)=>{
     }
 }
 
-exports.deleteSection = async (req,res) => {
-    try {
-        const {sectionId} = req.params;
-        
-        await Section.findByIdAndDelete(sectionId);
+exports.deleteSection = async (req, res) => {
+	try {
 
-        // also need to delete from Course model
+		const { sectionId, courseId }  = req.body;
+		await Course.findByIdAndUpdate(courseId, {
+			$pull: {
+				courseContent: sectionId,
+			}
+		})
+		const section = await Section.findById(sectionId);
+		console.log(sectionId, courseId);
+		if(!section) {
+			return res.status(404).json({
+				success:false,
+				message:"Section not Found",
+			})
+		}
 
-        return res.status(200).json({
-            success:true,
-            message:"Section deleted successfully"
-        })  
-    } catch (error) {
-        return res.status(500).json({
-            success:false,
-            message:"Section delted successfully!"
-        })  
-    }
-}
+		//delete sub section
+		await SubSection.deleteMany({_id: {$in: section.subSection}});
+
+		await Section.findByIdAndDelete(sectionId);
+
+		//find the updated course and return 
+		const course = await Course.findById(courseId).populate({
+			path:"courseContent",
+			populate: {
+				path: "subSection"
+			}
+		})
+		.exec();
+
+		res.status(200).json({
+			success:true,
+			message:"Section deleted",
+			data:course
+		});
+	} catch (error) {
+		console.error("Error deleting section:", error);
+		res.status(500).json({
+			success: false,
+			message: "Internal server error",
+		});
+	}
+};
